% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{constituencies}
\alias{constituencies}
\title{Get constituencies}
\usage{
constituencies()
}
\value{
Returns a \code{{sf}} object for the danish constituencies
}
\description{
Get constituencies
}
\seealso{
Other data retrieval: 
\code{\link{get_geo}()},
\code{\link{municipalities}()},
\code{\link{parishes}()},
\code{\link{regions}()}
}
\concept{data retrieval}
