% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_con}
\alias{geo_con}
\title{Danish constituencies}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_con
}
\description{
Danish constituencies
}
\keyword{datasets}
