% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geo_get_spatial}
\alias{geo_get_spatial}
\title{Get Danish spatial data}
\usage{
geo_get_spatial(spatial)
}
\arguments{
\item{spatial}{one of "Danish zip codes", "Danish regions", "Danish police districts",
"Danish placenames - points", "Danish placenames - lines",
"Danish placenames - areas", "Danish parishes", "Danish municipalities",
"Danish jurisdictions" or "Danish constituencies"}
}
\value{
an object of either SpatialLinesDataFrame, SpatialPointsDataFrame or
    SpatialPolygonsDataFrame
}
\description{
Get Danish spatial data
}
\examples{
library(geoDK)
library(ggplot2)

dk_regions <- geo_get_spatial("Danish regions")
dk_placenames_points <- geo_get_spatial("Danish placenames - points")

df_regions <- make_tidy_poly(dk_regions)
long_barrows <- subset(dk_placenames_points, dk_placenames_points$feat_type == "langdysse")

ggplot() +
  geom_polygon(data = df_regions, aes(x = lng, y = lat, group = area), fill = "white",
               color = "black", show.legend = FALSE) +
  geom_point(data = tibble::as_tibble(long_barrows@coords), aes(x = coords.x1, y = coords.x2),
             color = "red") +
  coord_cartesian() +
  theme_minimal()

}
