% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_jur}
\alias{geo_jur}
\title{Danish jurisdictions}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_jur
}
\description{
Danish jurisdictions
}
\keyword{datasets}
