% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_mun}
\alias{geo_mun}
\title{Danish municipalities}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_mun
}
\description{
Danish municipalities
}
\keyword{datasets}
