% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_par}
\alias{geo_par}
\title{Danish parishes}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_par
}
\description{
Danish parishes
}
\keyword{datasets}
