% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_pol}
\alias{geo_pol}
\title{Danish police districts}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_pol
}
\description{
Danish police districts
}
\keyword{datasets}
