% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_reg}
\alias{geo_reg}
\title{Danish regions}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_reg
}
\description{
Danish regions
}
\keyword{datasets}
