% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geo_url_builder}
\alias{geo_url_builder}
\title{Helper function to build the urls}
\usage{
geo_url_builder(...)
}
\arguments{
\item{...}{insert the relevant parameters for the call.}
}
\value{
a url
}
\description{
The function builds the url needed in the geo_getter() function. It is not
exported.
}
\examples{
# Set login and password ----
login <- 'xxx'
password <- 'yyy'

# Build url
geo_url_builder(method = "adresse", vejnavn = "Rentemestervej", postnr = "2400", hits = "50", login = login, password = password)

}
