% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_zip}
\alias{geo_zip}
\title{Danish zip codes}
\format{SpatialPolygonsDataFrame}
\source{
\url{https://kortforsyningen.dk/}
}
\usage{
geo_zip
}
\description{
Danish zip codes
}
\keyword{datasets}
