% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodk_enrich.R
\name{geodk_enrich}
\alias{geodk_enrich}
\title{Enrich a dataframe with geographic data}
\usage{
geodk_enrich(x)
}
\arguments{
\item{x}{Dataframe. Preferably from \code{{dkstat}}}
}
\value{
Returns a dataframe of the dimensions \verb{cols +1 by rows}.
}
\description{
This function makes it easy for you to enrich your dataframe with appropriate geographical information. The main usage, right now at least, is to enrich statistical data from \code{{geodk}} with geometries for the appropriate levels.
}
