% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{get_geo}
\alias{get_geo}
\title{Get geogrpahies}
\usage{
get_geo(section = "postnumre")
}
\arguments{
\item{section}{The geographic level wanted.}
}
\value{
Returns a \code{{sf}} object for the specified geography
}
\description{
Get geogrpahies
}
\seealso{
Other data retrieval: 
\code{\link{constituencies}()},
\code{\link{municipalities}()},
\code{\link{parishes}()},
\code{\link{regions}()}
}
\concept{data retrieval}
