% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{Get all available levels}
\usage{
get_levels()
}
\value{
This funciton returns all available levels for the \verb{plot_*}
functions.
}
\description{
Get all available levels
}
\examples{
get_levels()

}
