% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_tidy.R
\name{make_tidy_poly}
\alias{make_tidy_poly}
\title{Turns a spatial object into a tidy tibble}
\usage{
make_tidy_poly(spatial)
}
\arguments{
\item{spatial}{an object of class SpatialPolygonsDataFrame}
}
\value{
tibble
}
\description{
This will then work with ggplot2
}
\examples{
library(geoDK)
library(ggplot2)

geo_reg <- geo_get_spatial("Danish regions")

df <- make_tidy_poly(geo_reg)

ggplot(df) +
  geom_polygon(aes(x = lng, y = lat, group = area, fill = regionkode),
               show.legend = FALSE) +
  coord_cartesian()
}
