% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{parishes}
\alias{parishes}
\title{Get parishes}
\usage{
parishes()
}
\value{
Returns a \code{{sf}} object for the danish parishes
}
\description{
Get parishes
}
\seealso{
Other data retrieval: 
\code{\link{constituencies}()},
\code{\link{get_geo}()},
\code{\link{municipalities}()},
\code{\link{regions}()}
}
\concept{data retrieval}
