% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_denmark.R
\name{plot_denmark}
\alias{plot_denmark}
\title{Plot a map of Denmark with given borders}
\usage{
plot_denmark(level = "regions", fill = NULL, color = NULL)
}
\arguments{
\item{level}{The desired administrative geographic level for the map. Should
be one of \code{get_levels()}}

\item{fill}{The desired color for the areas to be filled with. Could be
either a R-friendly color (name or hex code) or \code{names} to fill based on
the area names.}

\item{color}{The desired color for the area borders to be colored with. Could
be either a R-friendly color (name or hex code) or \code{names} to fill based on
the area names.}
}
\value{
Returns a ggplot object that can be manipulated as any other. The
object is automatically printed.
}
\description{
Plot a nice map of Denmark with borders at a given level. Options to fill
based on area names or color borders are included. A legend guide is shown if
less than 15 areas are plotted.
}
\examples{
plot_denmark()

plot_denmark(fill = "names")

}
