% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regions.R
\name{plot_regions}
\alias{plot_regions}
\title{Plot selected (or all) regions}
\usage{
plot_regions(
  region = c("Region Nordjylland", "Region Midtjylland", "Region Syddanmark",
    "Region Hovedstaden", "Region Sjælland")
)
}
\arguments{
\item{region}{Region(s) to plot}
}
\value{
Returns a \code{{ggplot2}} object and prints the plot as well.
}
\description{
Plot a vector of regions in Denmark. Just provide the name.
}
\examples{
plot_regions(region = c("Region Nordjylland", "Region Midtjylland"))
}
