% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefill_cache.R
\name{prefill_cache}
\alias{prefill_cache}
\title{Prefill data retrieval cache}
\usage{
prefill_cache(specific_section = NULL)
}
\arguments{
\item{specific_section}{Enter a seciton name (or vector of section names) you
would like to have pre-filled. Options can be found in
\code{dawaR::avaialable_sections()}}
}
\value{
Returns nothing. Is called for the side effect of prefilling the
cache.
}
\description{
This function prefills the cache of the underlying API-function.
By default the sections shown in the example below is prefilled. This is
meant to replace the cached dataset that once was in the package. It may be
useful to prefill the cache before using \code{geodk_enrich}.
}
\examples{
sections <- c(
  "regioner",
  "kommuner",
  "postnumre",
  "sogne",
  "regioner",
  "opstillingskredse",
  "afstemningsomraader"
)
}
