% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Danish regions}
\format{
\subsection{\code{regions}}{

An object of classes \code{sf} and \code{data.frame} with 5 rows and 8 columns:
\describe{
\item{code}{Unique code}
\item{name}{The Danish name of the region}
\item{nuts2}{EU nuts division}
\item{source_changed}{When was the source last changed on the data-provider's side}
\item{visualcenter_x, visualcenter_y}{Visual center of the polygon in WGS84}
\item{geometry}{Polygons for each region}
\item{last_update}{When was the data in the package last updated by maintainers}
}
}
}
\source{
Provided by The Danish Agency of Climate Data
https://dawadocs.dataforsyningen.dk/dok/om#vilkaar. Terms and conditions
apply.
}
\usage{
regions
}
\description{
A \code{sf} dataset providing geometry for the danish regions
}
\keyword{datasets}
