% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{municipalities}
\alias{municipalities}
\title{Get municipalities}
\usage{
municipalities()
}
\value{
Returns a \code{{sf}} object for the danish municipalities
}
\description{
Get municipalities
}
\seealso{
Other data retrieval: 
\code{\link{constituencies}()},
\code{\link{get_geo}()},
\code{\link{parishes}()},
\code{\link{regions}()}
}
\concept{data retrieval}
