% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{regions}
\alias{regions}
\title{Get regions}
\usage{
regions()
}
\value{
Returns a \code{{sf}} object for the danish regions
}
\description{
Get regions
}
\seealso{
Other data retrieval: 
\code{\link{constituencies}()},
\code{\link{get_geo}()},
\code{\link{municipalities}()},
\code{\link{parishes}()}
}
\concept{data retrieval}
